Large reorg and cleanup
Fixes some issues with interpolated values in regex as well.
This will set up for some follow up PR's to add tests and corrections
for date parsing issues, as well as mypy integration.

Removed lots of code that we don't use a production env, mostly
around evaluation of performance and some more proprietary features
like the Standford NER.
