+++ .pre-commit-config.yaml
@@ -10,3 +10,4 @@ repos:
     rev: 5.4.2
     hooks:
       - id: isort
+        args: [ --profile, black ]
