+++ philter_lite/__init__.py
@@ -1,4 +1,9 @@
 __version__ = "1.0.0"
 
-import philter_lite.coordinate_map
-import philter_lite.philter
+from philter_lite.coordinate_map import CoordinateMap
+from philter_lite.filters import Filter, filter_from_dict, load_filters
+from philter_lite.philter import detect_phi
+
+# Writers
+from .asterisk import transform_text_asterisk
+from .i2b2 import transform_text_i2b2
