+++ tests/test_philter.py
@@ -2,7 +2,8 @@ import os
 
 import pytest
 
-from philter_lite import philter
+import philter_lite
+from philter_lite import detect_phi, filter_from_dict, filters, load_filters
 
 
 def test_filter_from_dict():
@@ -15,14 +16,14 @@ def test_filter_from_dict():
         "phi_type": "SOMETHING",
     }
 
-    filter = philter.filter_from_dict(filter_dict)
+    filter = filter_from_dict(filter_dict)
 
     assert filter.type == "regex"
     assert filter.title == "test_city"
     assert filter.data is not None
     assert filter.exclude == "test_ex"
     assert filter.phi_type == "SOMETHING"
-    assert isinstance(filter, philter.RegexFilter)
+    assert isinstance(filter, filters.RegexFilter)
 
     filter_dict = {
         "title": "Find Names 1",
@@ -34,7 +35,7 @@ def test_filter_from_dict():
         "phi_type": "Something",
     }
 
-    filter = philter.filter_from_dict(filter_dict)
+    filter = filter_from_dict(filter_dict)
 
     assert filter.type == "regex_context"
     assert filter.title == "Find Names 1"
@@ -42,7 +43,7 @@ def test_filter_from_dict():
     assert filter.exclude is True
     assert filter.context == "right"
     assert filter.context_filter == "Firstnames Blacklist"
-    assert isinstance(filter, philter.RegexContextFilter)
+    assert isinstance(filter, filters.RegexContextFilter)
 
     filter_dict = {
         "title": "Whitelist 1",
@@ -53,14 +54,14 @@ def test_filter_from_dict():
         "phi_type": "Something",
     }
 
-    filter = philter.filter_from_dict(filter_dict)
+    filter = filter_from_dict(filter_dict)
 
     assert filter.type == "set"
     assert filter.title == "Whitelist 1"
     assert filter.data is not None
     assert filter.exclude is False
     assert filter.pos == []
-    assert isinstance(filter, philter.SetFilter)
+    assert isinstance(filter, filters.SetFilter)
 
     filter_dict = {
         "title": "POS MATCHER",
@@ -70,13 +71,13 @@ def test_filter_from_dict():
         "phi_type": "OTHER",
     }
 
-    filter = philter.filter_from_dict(filter_dict)
+    filter = filter_from_dict(filter_dict)
 
     assert filter.type == "pos_matcher"
     assert filter.title == "POS MATCHER"
     assert filter.exclude is False
     assert filter.pos == ["CD"]
-    assert isinstance(filter, philter.PosFilter)
+    assert isinstance(filter, filters.PosFilter)
 
 
 def test_filter_from_dict_missing_phi_type():
@@ -88,8 +89,8 @@ def test_filter_from_dict_missing_phi_type():
         "notes": "test_notes",
     }
 
-    with pytest.raises(KeyError):
-        philter.filter_from_dict(filter_dict)
+    filter = filter_from_dict(filter_dict)
+    assert filter.phi_type == "OTHER"
 
 
 def test_filter_from_dict_missing_file():
@@ -99,12 +100,21 @@ def test_filter_from_dict_missing_file():
     }
 
     with pytest.raises(Exception):
-        philter.filter_from_dict(filter_dict)
+        filter_from_dict(filter_dict)
 
 
-def test_map_coordinates():
+def test_default_config():
+    filters = load_filters(
+        os.path.join(
+            os.path.dirname(philter_lite.__file__), "configs/philter_delta.toml"
+        )
+    )
+    assert len(filters) > 0
+
+
+def test_detect_phi():
     patterns = [
-        philter.filter_from_dict(
+        filter_from_dict(
             {
                 "title": "patient SSN",
                 "type": "regex",
@@ -114,7 +124,7 @@ def test_map_coordinates():
                 "phi_type": "MRN",
             }
         ),
-        philter.filter_from_dict(
+        filter_from_dict(
             {
                 "title": "dd_mm_yyyy",
                 "type": "regex",
@@ -125,10 +135,35 @@ def test_map_coordinates():
             }
         ),
     ]
-    text_data, include_map, exclude_map, data_tracker = philter.map_coordinates(
-        "The patients SSN is 123-45-6789. They were born on 01/01/1984",
+    include_map, exclude_map, data_tracker = detect_phi(
+        "The patients SSN is 123-45-6789. They were born on 01/01/1984.",
         patterns,
         ["MRN", "DATE"],
     )
 
     assert len(data_tracker.phi) == 2
+
+
+def test_detect_phi_regex_interpolation():
+    # At runtime, we compute the seasons value in the regex, by interpolating a variable name.
+    patterns = [
+        filter_from_dict(
+            {
+                "title": "season of yyyy",
+                "type": "regex",
+                "exclude": True,
+                "notes": "",
+                "phi_type": "DATE",
+                "keyword": "dates.season_of_yyyy",
+            }
+        ),
+    ]
+    include_map, exclude_map, data_tracker = detect_phi(
+        "They injured themselves in Fall of 2020.", patterns, ["MRN", "DATE"],
+    )
+
+    assert len(data_tracker.phi) == 1
+    assert data_tracker.phi[0].start == 27
+    assert data_tracker.phi[0].stop == 39
+    assert data_tracker.phi[0].word == "Fall of 2020"
+    assert data_tracker.phi[0].phi_type == "DATE"
